refactor: 将图标逻辑内联到处理器类
将原先独立的 addIconToPluginNavItem 工具函数移除，并
将其实现内联到 CommunityPluginIconHandler 类中作为私有
方法 addIconToPluginNavItem。与此同时引入 setIcon 直接用
于更新图标，删除对外部 util 的依赖。

这样做的原因是将与插件图标处理紧密相关的逻辑封装在
处理器内部，减少模块间耦合，便于访问实例上下文和设
置，并且简化模块结构。保留了对原生图标的跳过判断、
自定义图标容器的创建/插入、按需更新以及数据属性的
缓存，确保行为不变并提高可维护性。
